refactor: isolate ambient authority in index.js
